# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "restaurantAPI",
        "version": "0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Осуществляет вход пользователя по логину\\паролю или пин-коду",
                "operationId": "login",
                "responses": {
                    "200": {
                        "description": "Произведен вход в систему"
                    },
                    "400": {
                        "description": "Неверный логин\\пароль или пин-код"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Выводит категории блюд",
                "operationId": "categoriesAll",
                "responses": {
                    "200": {
                        "description": "Категории выведены"
                    },
                    "404": {
                        "description": "Категории не найдены"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Создает категорию",
                "operationId": "categoriesCreate",
                "responses": {
                    "200": {
                        "description": "Категория создана"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/categories/{category_id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Вывод категории по id",
                "operationId": "categoryGetById",
                "responses": {
                    "200": {
                        "description": "Продукт выведен"
                    },
                    "404": {
                        "description": "Продукт не найден"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Обновляет категорию",
                "operationId": "categoriesUpdate",
                "responses": {
                    "200": {
                        "description": "Категория обновлена"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Удаляет категорию",
                "operationId": "categoriesDelete",
                "responses": {
                    "200": {
                        "description": "Категория удалена"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/dishes": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Вывод всех блюд",
                "operationId": "dishesGetAll",
                "responses": {
                    "200": {
                        "description": "Блюда выведены"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Создает блюдо",
                "operationId": "dishesCreate",
                "responses": {
                    "200": {
                        "description": "Блюдо создано"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/dishes/{dish_id}": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Вывод блюда по id",
                "operationId": "dishesGetById",
                "responses": {
                    "200": {
                        "description": "Блюдо выведено"
                    },
                    "404": {
                        "description": "Блюдо не найдено"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/dishes/{dishes_id}": {
            "put": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Обновляет блюдо по id",
                "operationId": "dishesUpdate",
                "responses": {
                    "200": {
                        "description": "Блюдо обновлено"
                    },
                    "404": {
                        "description": "Блюдо не найдено"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Удаляет блюдо по id",
                "operationId": "dishesDelete",
                "responses": {
                    "200": {
                        "description": "Блюдо удалено"
                    },
                    "404": {
                        "description": "Блюдо не найдено"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
...
    }
}